Fix device placement and tensor dtype handling for ControlNet depth mode
Key fixes:
- Fixed hardcoded 'cuda:1' device references to use self.device consistently
- Fixed tensor device mismatch in ControlNet pipeline prepare_image method by explicitly converting to float32 before dtype conversion
- Simplified control image preprocessing to keep PIL images until pipeline processing
- Updated Dockerfile to attempt Depth Anything V2 installation (gracefully falls back to fallback methods)
- Improved error handling for missing Depth Anything V2 and AnylineDetector modules with graceful fallbacks
- Updated protobuf requirement to >=4.25.3 for compatibility

Changes:
- model.py: Fixed device handling in control image processing and text embedding computation
- flux/pipeline_flux_controlnet_img2img.py: Added explicit float() conversion in prepare_image
- requirements.txt: Updated protobuf version constraint
- Dockerfile: Added attempt to install Depth Anything V2 with fallback

+++ flux/pipeline_flux_controlnet_img2img.py
@@ -589,7 +589,8 @@ class FluxControlNetImg2ImgPipeline(DiffusionPipeline, FluxLoraLoaderMixin, From
 
         image = image.repeat_interleave(repeat_by, dim=0)
 
-        image = image.to(device=device, dtype=dtype)
+        # Ensure image is on correct device and dtype
+        image = image.float().to(device=device, dtype=dtype)
 
         if do_classifier_free_guidance and not guess_mode:
             image = torch.cat([image] * 2)
